refactor(tests): ♻️ split cross-platform tests per platform
- inline test_all_entities into each test_<platform>.py (was parametrized in test_entities.py)
- inline test_setup_when_modules_absent into each test_<platform>.py (was test_skip_branches.py)
- split tests/snapshots/test_entities.ambr into 8 per-platform snapshot files
- delete tests/test_entities.py, tests/test_skip_branches.py, tests/snapshots/test_entities.ambr
- side benefit: time platform now covered by test_all_entities (was missing from old cross-platform parametrize)

293 tests pass, 100% coverage retained.
