+++ tests/snapshots/test_light.ambr
@@ -1,0 +1,48 @@
+# serializer version: 1
+# name: test_all_entities
+  list([
+    EntityRegistryEntrySnapshot({
+      'aliases': list([
+        None,
+      ]),
+      'area_id': None,
+      'capabilities': dict({
+        'supported_color_modes': list([
+          <ColorMode.ONOFF: 'onoff'>,
+        ]),
+      }),
+      'config_entry_id': <ANY>,
+      'config_subentry_id': <ANY>,
+      'device_class': None,
+      'device_id': <ANY>,
+      'disabled_by': None,
+      'domain': 'light',
+      'entity_category': None,
+      'entity_id': 'light.pool_pool_light',
+      'has_entity_name': True,
+      'hidden_by': None,
+      'icon': None,
+      'id': <ANY>,
+      'labels': set({
+      }),
+      'name': None,
+      'object_id_base': 'Pool Light',
+      'options': dict({
+      }),
+      'original_device_class': None,
+      'original_icon': None,
+      'original_name': 'Pool Light',
+      'platform': 'neopool',
+      'previous_unique_id': None,
+      'suggested_object_id': None,
+      'supported_features': 0,
+      'translation_key': 'light',
+      'unique_id': 'neopool_1234567890_light',
+      'unit_of_measurement': None,
+    }),
+  ])
+# ---
+# name: test_setup_when_modules_absent
+  list([
+  ])
+# ---
