--- tests/test_entities.py
@@ -1,53 +1,0 @@
-"""Snapshot the entities created by every neopool platform."""
-
-from unittest.mock import MagicMock, patch
-
-import pytest
-from pytest_homeassistant_custom_component.common import MockConfigEntry
-from syrupy.assertion import SnapshotAssertion
-
-from homeassistant.const import Platform
-from homeassistant.core import HomeAssistant
-from homeassistant.helpers import entity_registry as er
-
-from . import setup_integration
-
-
-@pytest.mark.parametrize(
-    "platform",
-    [
-        Platform.BINARY_SENSOR,
-        Platform.BUTTON,
-        Platform.LIGHT,
-        Platform.NUMBER,
-        Platform.SELECT,
-        Platform.SENSOR,
-        Platform.SWITCH,
-    ],
-)
-async def test_all_entities(
-    hass: HomeAssistant,
-    snapshot: SnapshotAssertion,
-    entity_registry: er.EntityRegistry,
-    mock_config_entry: MockConfigEntry,
-    mock_neopool_client: MagicMock,
-    platform: Platform,
-) -> None:
-    """Snapshot the entities each platform registers from the default fixture.
-
-    Snapshot the registry entries directly rather than via
-    `snapshot_platform`, which assumes every entity is enabled and has
-    state. NeoPool ships several `entity_registry_enabled_default=False`
-    entities (e.g. MBF_CELL_BOOST), and including them via state lookup
-    would either fail or pull entire state machines into the snapshot.
-    The registry entry alone (unique_id, name, disabled_by, …) is the
-    stable shape we care about.
-    """
-    with patch("custom_components.neopool.PLATFORMS", [platform]):
-        await setup_integration(hass, mock_config_entry)
-
-    entries = sorted(
-        er.async_entries_for_config_entry(entity_registry, mock_config_entry.entry_id),
-        key=lambda e: e.entity_id,
-    )
-    assert entries == snapshot
