+++ tests/test_light.py
@@ -1,9 +1,10 @@
 """Tests for the NeoPool light platform."""
 
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, patch
 
 import pytest
 from pytest_homeassistant_custom_component.common import MockConfigEntry
+from syrupy.assertion import SnapshotAssertion
 
 from custom_components.neopool.light import LIGHT_DESCRIPTIONS
 from homeassistant.components.light import DOMAIN as LIGHT_DOMAIN
@@ -13,6 +14,7 @@ from homeassistant.const import (
     STATE_OFF,
     STATE_ON,
     STATE_UNAVAILABLE,
+    Platform,
 )
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_platform as ep, entity_registry as er
@@ -155,3 +157,56 @@ async def test_light_winter_mode_guard_when_called_directly(
     await entity_obj.async_turn_off()
     assert "Winter mode is active" in caplog.text
     mock_neopool_client.async_write_register.assert_not_called()
+
+
+# ---------------------------------------------------------------------------
+# Platform-wide snapshots
+# ---------------------------------------------------------------------------
+
+
+async def test_all_entities(
+    hass: HomeAssistant,
+    snapshot: SnapshotAssertion,
+    entity_registry: er.EntityRegistry,
+    mock_config_entry: MockConfigEntry,
+    mock_neopool_client: MagicMock,
+) -> None:
+    """Snapshot every entity registered by the light platform.
+
+    Snapshot the registry entries directly rather than via
+    `snapshot_platform`, which assumes every entity is enabled and has
+    state. NeoPool ships several `entity_registry_enabled_default=False`
+    entities; including them via state lookup would either fail or pull
+    entire state machines into the snapshot. The registry entry alone
+    (unique_id, name, disabled_by, ...) is the stable shape we care about.
+    """
+    with patch("custom_components.neopool.PLATFORMS", [Platform.LIGHT]):
+        await setup_integration(hass, mock_config_entry)
+    entries = sorted(
+        er.async_entries_for_config_entry(entity_registry, mock_config_entry.entry_id),
+        key=lambda e: e.entity_id,
+    )
+    assert entries == snapshot
+
+
+async def test_setup_when_modules_absent(
+    hass: HomeAssistant,
+    snapshot: SnapshotAssertion,
+    entity_registry: er.EntityRegistry,
+    mock_config_entry: MockConfigEntry,
+    mock_neopool_client_minimal: MagicMock,
+) -> None:
+    """Snapshot the light entities registered when no modules are present.
+
+    Drives setup with the lean `mock_neopool_client_minimal` fixture (no
+    modules detected, no relay GPIOs assigned). Each platform's gating
+    branches fire and entities depending on the missing hardware are
+    skipped; the resulting registry shape is captured as a snapshot.
+    """
+    with patch("custom_components.neopool.PLATFORMS", [Platform.LIGHT]):
+        await setup_integration(hass, mock_config_entry)
+    entries = sorted(
+        er.async_entries_for_config_entry(entity_registry, mock_config_entry.entry_id),
+        key=lambda e: e.entity_id,
+    )
+    assert entries == snapshot
